+++ packages/partitura_core/lib/partitura_core.dart
@@ -45,5 +45,6 @@ export 'src/theory/key_signature.dart';
 export 'src/theory/pitch.dart';
 export 'src/theory/scale.dart';
 export 'src/theory/time_signature.dart';
+export 'src/theory/transposition.dart';
 export 'src/theory/triad.dart';
 export 'src/theory/tuning.dart';